backend/requirements.txt
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-generator==1.10
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
blis==1.3.3
build==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.15
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
distro==1.9.0
dnspython==2.8.0
docx2txt==0.9
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
faiss-cpu==1.12.0
fastapi==0.121.0
ffmpeg-python==0.2.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.9.23
flatten-dict==0.4.2
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.0.4
langchain-core==1.0.3
langcodes==3.5.1
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.10.2
networkx==3.5
numba==0.62.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.20.0
openai-whisper==20250625
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pefile==2024.8.26
pillow==12.0.0
posthog==5.4.0
preshed==3.0.12
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybase64==1.4.2
pycparser==3.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
PyMuPDF==1.26.6
PyMySQL==1.1.2
pyparsing==3.2.5
pypdf==6.7.0
PyPDF2==3.0.1
pyperclip==1.11.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
selenium==4.40.0
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.46
srsly==2.5.2
starlette==0.49.3
streamlit==1.51.0
streamlit-audiorec @ git+https://github.com/stefanrmmr/streamlit-audio-recorder.git@777d18114130137d492c0378a86631fff1ff1be5
streamlit_mic_recorder==0.0.8
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
timeout-decorator==0.5.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.1
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.38.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
wsproto==1.3.2
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0

groq
